+++ presentation/src/main/java/com/threegap/bitnagil/presentation/guide/contract/GuideState.kt
@@ -1,4 +1,6 @@
-package com.threegap.bitnagil.presentation.guide.model
+package com.threegap.bitnagil.presentation.guide.contract
+
+import com.threegap.bitnagil.presentation.guide.model.GuideType
 
 data class GuideState(
     val guideType: GuideType?,
